Move command-char tests to a file CI executes
run-crash-handler.test.ts is skip-listed in test/expectations.txt (its
segfault-reporter test is broken), so tests added there never run in
CI. Move the crash-report command character tests to their own file.
